Move wavelength UI from controls to tools
